+++ tests/io/test_net.py
@@ -9,7 +9,7 @@ import pytest
 from meshkernel import DeleteMeshOption, GeometryList, MeshKernel
 
 from hydrolib.core.basemodel import BaseModel
-from hydrolib.core.io.mdu.models import FMModel
+from hydrolib.core.io.dflowfm.mdu.models import FMModel
 from hydrolib.core.io.net.models import Branch, Mesh2d, NetworkModel
 from hydrolib.core.io.net.reader import NCExplorer
 from hydrolib.core.io.net.writer import FillValueConfiguration, UgridWriter
