+++ tests/test_utils.py
@@ -2,7 +2,7 @@ from pathlib import Path
 
 import pytest
 
-from hydrolib.core.io.mdu.models import Geometry, Output
+from hydrolib.core.io.dflowfm.mdu.models import Geometry, Output
 from hydrolib.core.utils import get_substring_between, str_is_empty_or_none
 
 from .utils import test_input_dir
